+++ src/FSharp.Data.GraphQL.Shared/TypeSystem.fs
@@ -809,9 +809,9 @@ and ExecutionInfoKind =
 /// Buffered stream options. Used to specify how the buffer will behavior in a stream.
 and BufferedStreamOptions = {
     /// The maximum time in milliseconds that the buffer will be filled before being sent to the subscriber.
-    Interval : int option
+    Interval : int voption
     /// The maximum number of items that will be buffered before being sent to the subscriber.
-    PreferredBatchSize : int option
+    PreferredBatchSize : int voption
 }
 
 /// <summary>